Redirect and on click operations isolated
Changed the ticker's onClick to onDoubleClick. When it's onClick it performs both the redirect and brings up the sell prompt at the same time (sell prompt appears at the redirected market page when a sell is not intended). onDoubleClick isolates the two operations to do one or the other and not both together.